Fix Monokai yellow
The yellow color was pretty far off from what I found in the actual
Monokai colorscheme. It was set to f4/bf/75, which is identical to
base16-default's yellow. I'm betting that this color was accidentally
missed when the Monokai version was created. In this commit I am
changing it to the yellow that I found in the Monokai colorscheme that
ships with Sublime Text 2.