Rename some themes to match latest schemas
In
https://github.com/chriskempson/base16-vim/commit/a993579e#diff-e2ad1445e89d0943e05d06fda12f8d4b
these were changed, but the shell repo was not updated so these themes
ended up breaking. This commit corrects this for a few that I spotted.

There seem to be some more changes in the commit I linked above (e.g.
some new themes maybe), but this is probably a good start.