Fix solarized light
In eeeff9d solarized.light was removed, and only solarized was present.
It was re-added in 128d28b, but with different color values than before
eeeff9d.

This makes it use the previous values again, which are the same as with
the dark variant, except for fg/bg (and cursor, but that is unused).