Fix BASE16_THEME env set by fish shell helper
The fish shell helper script was using the wrong variable to parse the
current theme name from the theme's path, thus setting the BASE16_THEME
environment variable to an empty string.

This affects scripts or configurations that rely on BASE16_THEME, such
as the base16-vim users that use the instructions on the README file
(colorschema base16-$BASE16_THEME).